[BBB-136] ✨Feat: circuit breaker를 사용하여 Redis 장애에도 정상적으로 로그인 암복호화가 진행되도록 적용
Redis 장애 시, 이미 인메모리에 저장된 서버 간에 공유되는 비대칭키를 사용하도록 하여 항상 로그인 요청이 처리될 수 있도록 보완했습니다.

+++ app/external-api/src/main/java/com/bombombom/devs/external/encryption/controller/dto/ProvidePublicKeyResponse.java
@@ -6,7 +6,7 @@ import lombok.Builder;
 @Builder
 public record ProvidePublicKeyResponse(
     int id,
-    Double version,
+    long version,
     String publicKey
 ) {
