+++ app/external-api/src/main/java/com/bombombom/devs/external/global/decryption/dto/EncryptedRequest.java
@@ -2,7 +2,7 @@ package com.bombombom.devs.external.global.decryption.dto;
 
 public record EncryptedRequest(
     int id,
-    double version,
+    long version,
     byte[] encryptedData
 ) {
